+++ libmproxy/protocol/__init__.py
@@ -1,1 +1,12 @@
-from .primitives import *
+from __future__ import (absolute_import, print_function, division)
+from .base import Layer, ServerConnectionMixin, Log, Kill
+from .http import Http1Layer, Http2Layer
+from .tls import TlsLayer, is_tls_record_magic
+from .rawtcp import RawTCPLayer
+
+__all__ = [
+    "Layer", "ServerConnectionMixin", "Log", "Kill",
+    "Http1Layer", "Http2Layer",
+    "TlsLayer", "is_tls_record_magic",
+    "RawTCPLayer"
+]
