--- libmproxy/protocol2/__init__.py
@@ -1,13 +1,0 @@
-from __future__ import (absolute_import, print_function, division)
-from .root_context import RootContext
-from .socks_proxy import Socks5Proxy
-from .reverse_proxy import ReverseProxy
-from .http_proxy import HttpProxy, HttpUpstreamProxy
-from .transparent_proxy import TransparentProxy
-from .http import make_error_response
-
-__all__ = [
-    "RootContext",
-    "Socks5Proxy", "ReverseProxy", "HttpProxy", "HttpUpstreamProxy", "TransparentProxy",
-    "make_error_response"
-]
