+++ libmproxy/proxy/modes/socks_proxy.py
@@ -2,8 +2,9 @@ from __future__ import (absolute_import, print_function, division)
 
 from netlib import socks
 from netlib.tcp import NetLibError
-from ..exceptions import Socks5Exception
-from .layer import Layer, ServerConnectionMixin
+
+from ...exceptions import Socks5Exception
+from ...protocol import Layer, ServerConnectionMixin
 
 
 class Socks5Proxy(Layer, ServerConnectionMixin):
